requirements-test.txt
eth_keys
eth-account>=0.4.0,<0.6.0
mpmath==1.0.0
pytest>=7.0.0
six>=1.14
python-dotenv==0.17.1
web3>=5.0.0,<6.0.0
